Adapt most tests to new framework
