+++ src/platform/test/test-main.cpp
@@ -21,28 +21,9 @@
 #include "test-functions.hpp"
 #include <ariajanke/cul/TreeTestSuite.hpp>
 
-#include <array>
-#include <algorithm>
-
-#include <vector>
-#include <memory>
-#include <atomic>
-
 int main() {
-    // I *could* just dump all my test functions here I guess...
-    const std::array k_test_functions = {
-        run_spm_tests,
-        run_triangle_segment_tests,
-        run_map_loader_tests,
-        run_triangle_links_tests,
-        run_systems_tests,
-        run_wall_tile_factory_tests,
-    };
-    std::array<bool, k_test_functions.size()> results;
-    for (std::size_t i = 0; i != k_test_functions.size(); ++i)
-        results[i] = k_test_functions[i]();
-    bool all_ok = std::all_of(results.begin(), results.end(),
-                [](bool b){ return b; });
-    cul::tree_ts::run_tests();
-    return all_ok ? 0 : ~0;
+    // keeping these legacy tests
+    if (!run_systems_tests()) return ~0;
+
+    return cul::tree_ts::run_tests();
 }
