+++ tests/TaskController-tests.cpp
@@ -19,6 +19,7 @@
 *****************************************************************************/
 
 #include "../src/TasksController.hpp"
+// forward declarations
 #include "../src/PointAndPlaneDriver.hpp"
 
 #include <ariajanke/cul/TreeTestSuite.hpp>
